webserver/requirements.txt
# Pip requirements for webserver
flask
numpy
PyTurboJPEG